Remove dependency on ContractUnlocker in inputs and remove unnecessary methdos from Advanced builder
