+++ packages/cashscript/src/advanced/constants.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_SEQUENCE = 0xfffffffe;
+export const DEFAULT_LOCKTIME = 0;
