+++ packages/cashscript/src/advanced/interfaces.ts
@@ -1,10 +1,10 @@
 import type SignatureTemplate from '../SignatureTemplate.js';
-import type { Contract, ContractUnlocker } from '../Contract.js';
+import type { Contract } from '../Contract.js';
 import { Utxo, Unlocker } from '../interfaces.js';
 import { NetworkProvider } from '../network/index.js';
 
 export interface UnlockableUtxo extends Utxo {
-  unlocker: ContractUnlocker | Unlocker;
+  unlocker: Unlocker | ((...args: any[]) => Unlocker);
   options?: InputOptions;
 }
